third_party/WebKit/LayoutTests/clipboard/readtext-denied.https.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../http/tests/resources/permissions-helper.js"></script>
</head>
<body>
<script>
description("Tests navigator.clipboard.readText() permission failure.");

shouldBeDefined("PermissionsHelper.setPermission");

function successCallback(access) {
  testFailed(
    "clipboard.readText() does not fail (even though it really should have).");
  finishJSTest();
}

function errorCallback(error) {
  testPassed("clipboard.readText() fail (as expected).");
  finishJSTest();
}

window.jsTestIsAsync = true;

PermissionsHelper.setPermission('clipboard-read', 'denied').then(function() {
  shouldBeNonNull("navigator.clipboard");
  navigator.clipboard.readText().then(successCallback, errorCallback);
});


</script>
</body>
</html>
