third_party/WebKit/LayoutTests/clipboard/readtext-denied.https-expected.txt

Tests navigator.clipboard.readText() permission failure.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PermissionsHelper.setPermission is defined.
PASS navigator.clipboard is non-null.
PASS clipboard.readText() fail (as expected).
PASS successfullyParsed is true

TEST COMPLETE
